Classified but it need to be check.
@samsam2310  @lacy108

+++ codebook/code/Building Environment/default.cpp
@@ -1,6 +1,6 @@
 #define F(n) Fi(i,n)
 #define Fi(i,n) Fl(i,0,n)
-#define Fl(i,l,n) for(int i=l;i<n;++i)
+#define Fl(i,l,n) for(int i=(l);i<(int)(n);++i)
 #include <bits/stdc++.h>
 #include <bits/extc++.h>
 // #include <ext/pb_ds/assoc_container.hpp>
